[clean]: パッケージの整理
