+++ src/main/java/net/lifecity/mc/skillmaster/game/games/Duel.java
@@ -1,17 +1,20 @@
-package net.lifecity.mc.skillmaster.game.oneonone;
+package net.lifecity.mc.skillmaster.game.games;
 
-import net.lifecity.mc.skillmaster.game.*;
+import net.lifecity.mc.skillmaster.game.Game;
+import net.lifecity.mc.skillmaster.game.GameTeam;
+import net.lifecity.mc.skillmaster.game.GameType;
+import net.lifecity.mc.skillmaster.game.field.TwoPointField;
 import net.lifecity.mc.skillmaster.user.SkillUser;
 import org.bukkit.ChatColor;
 
-public class OneOnOne extends Game {
+public class Duel extends Game {
 
     private final TwoPointField field;
 
     private final GameTeam teamA;
     private final GameTeam teamB;
 
-    protected OneOnOne(TwoPointField field, SkillUser userA, SkillUser userB) {
+    protected Duel(TwoPointField field, SkillUser userA, SkillUser userB) {
         super(GameType.ONE_ON_ONE, 240, 6);
         this.field = field;
         this.teamA = new GameTeam("Alpha", ChatColor.RED, new SkillUser[]{userA});
@@ -20,15 +23,7 @@ public class OneOnOne extends Game {
 
     @Override
     public void sendResult() {
-        // 結果集計
-        String sub = "A:33 --- B:4";
 
-        GameTeam winner = teamA;
-        GameTeam loser = teamB;
-
-        // タイトル送信
-        winner.sendTitle(GameResult.WIN.getEn(), sub);
-        loser.sendTitle(GameResult.LOSE.getEn(), sub);
     }
 
     @Override
